+++ lib/vmpooler/statsd.rb
@@ -34,23 +34,4 @@ module Vmpooler
       $stderr.puts "Failure updating timing #{prefix}.#{label} on statsd server [#{server}:#{port}]: #{err}"
     end
   end
-
-  class DummyStatsd
-    attr_reader :server, :port, :prefix
-
-    def initialize(params = {})
-    end
-
-    def increment(label)
-      true
-    end
-
-    def gauge(label, value)
-      true
-    end
-
-    def timing(label, duration)
-      true
-    end
-  end
 end
